vastly simplify cursor loading by caching a default cursor if not found
this commit changes the behaviour of the "override" cursor to fall back to
the default cursor immediately, instead of trying to render the current
cursor status first. i don't know which behaviour is more intuitive,
but this is also how i implemented it in cosmic and it should not change
much for most use cases.

this also removes two very annoying usages of `unsafe` and an expect.