get input device info from mayland via mayctl devices
as devices are, similar to sway, split up into seperate devices, one
for each capability. all of those devices will then be serialized and
sent over the socket to the client.

unlike the default libinput capabilities, this will also distinguish
between a Touchpad and more the more generic Pointer, which is every
other pointer device.