add ability to reload config with mayctl reload
when invoking `mayctl reload`, and the config file is valid, then
mayland will read the new config, and update the things that changed.
this involves deriving PartialEq and Eq on all types from
`mayland-config`, to check if the config has changed, and to check what
actually needs to be updated.