hide winit cursor
when running mayland in winit mode, it will render two cursors: the one
of the "outside" compositor on top of the winit window, which will always
be the default cursor, and the one rendered by mayland. this isn't like
the end of the world, but it is still pretty annoying and hiding the
outside cursor is a single line change, so here it is.