support configuring output mode
similar to sway this picks the mode with the *closest* configured
refresh rate of the modes that match the configured size *exactly*.
if no refresh is given, it picks the highest.

if it couldn't match a configured size or no output config was given,
it tries to get the preferred mode with the highest quality.