yknow what, just always set the window state to tiled
i don't want to not render a focus ring for windows,
that are client side decorated, because they sometimes
render rounded corners, resulting in a gap between
the window border and the focus ring. but if i just
always tell the client that they are tiled, gtk windows
will fill the rectangle and not render rounded corners.