refactor and improve workspace handling
don't default back to the first workspace, when
the current workspace is requested and no active
output was found, and instead let the calling places
properly handle what to do, when no active workspace
exists.

this also makes some functions more consistent and
makes it clearer if a function is operating on all
workspaces, or just the active one.