make workspaces belong to a specific output
previously workspaces did not belong to a specific output, but would be
mapped to the currently active output on workspace switch, except when
they were already displayed on another output.

this changes workspace behaviour to be more intuitive and in line with
how other compositors do it by keeping track of which workspace belongs
on which output, and only mapping the workspace to that output instead of
the active one.

when an output is removed, all its workspaces become "orphans", ie their
outputs are removed and the next time they are mapped they will be mapped
to the then active output, or if another output is added again.

to make this properly work i also have to remove empty workspaces
(workspaces with no windows) from the workspace map, when they are no
longer being used.