fix workspace comm_info active field
i previously checked if the workspace is active, by checking if its
mapped output is the active output, but that is not possible with the new
workspace system, as workspaces don't lose their mapped outputs anymore,
when they are unfocussed.

instead i now get the active output, and check if it is the same as
`self` by checking if their indexes are the same.