add implicit ^(?:)$ to regex windowrules match
this makes it so you have to do an explicit `.*`, instead of this being
done implicitly. i think this is a fine tradeoff since i almost
exclusively use full regex matches, so i always have to add a `^$`
myself, and i think for the few times that i don't want that behaviour
it's fine (if not clearer) to add an explicit `.*`.