add implicit ^(?:)$ to regex windowrules match
this makes it so you have to do an explicit `.*`, instead of this being
done implicitly. i think this is a fine tradeoff since i almost
exclusively use full regex matches, so i always have to add a `^$`
myself, and i think for the few times that i don't want that behaviour
it's fine (if not clearer) to add an explicit `.*`.

this is not a final decision, especially since i already have regular
full string matches, and not every windowrule is a regex match like how
other compositors like hyprland do it, but i think it might be more
intuitive for it to match the entire string, so it is consistent with
how non-regex matchers work.