partially fix keyboard and pointer focus for layer shell surfaces
previously while if, when updating the keyboard focus, an overlay layer
was under the cursor, the keyboard would focus that layer, forget to
return and so it would then immediately override that focus by focussing
a window, if any window was mapped on that workspace, meaning that no
layer was ever able to actually get keyboard focus.

i also forgot to add the output and layer geometry to the return value of
surface_under.

this still does not cover layer surfaces with exclusive keyboard focus
and the code is still pretty bad, but that will have to be done another
time.