refactor layer shell focus
cleanup the surface_under and update_keyboard_focus functions by putting
shared logic for getting a layer surface in a `layer_surface_under` fn,
that takes a list of `Layer`s and checks for each of them if it finds a
layer surface under the given point.

the `layer_surface_under` fn takes an argument for if we're looking for
a `Pointer` or a `Keyboard` surface focus, where it will also check if
the surface can even receive keyboard focus, and skips it if not.