remove a few unnecessary From impls
the `impl From<UnmappedSurface> for Window` is significantly better as
an inherent `fn new` constructor, allowing for future extensibility like
passing in the computed windowrules. it also removes an unnecessary
indirection, because the previous `new` fn was only called from the
`From` impl and that one was called exactly once.

for the KeyboardFocusTarget and PointerFocusTarget some of the `From`
impls were unused already, and i could easily change at least one of the
others, cause i think that an explicit enum construction is just a lot
more clear than a `::from` call.