newtype Arc<RwLock<WindowRule>> and add getters
this is mostly for better and more readable code, unlocking a RwLock is
a little annoying for the "end user" (me), because you also always have
to `unwrap` the result to deal with poison, so a more simple getter just
makes the code better and more readable.

i previously tried to do that with a `fn windowrules` on MappedWindow,
that returned a `RwLockReadGuard`, but i just do not like returning
guards, because that has a whole slew of lifetime problems and potential
deadlocks, and i already have to always be careful when using the
smithay layer_map, because that one is also just a `MutexGuard`, and i
don't want another hazard like that around.