create a new Config::read fn, that returns a more proper Result
the previous `Config::read` fn is renamed to `Config::init`, and is the
one actually responsible for printing the pretty errors, and handling
a missing config. the `Config::read` fn, which is used inside of
`Config::init`, only returns an `Error` and doesn't print anything or
fallback to the default Config, when the file does not exist.

the `Config::read` fn will be used when reloading, where i don't want to
change back to the default config on deletion and i don't want to print
errors, but in the meantime it'll allow me to clean up some of the error
handling code.