don't default insert a Action::Quit bind when none is set
that was initially there so you couldn't accidentally lock yourself into
mayland, but apparently there is a use case for not wanting a quit
shortcut at all, so i just removed it.

it's probably fine because the default config provides a quit shortcut
and if you don't have that, you can always just start a terminal and use
mayctl to quit from there and if that isn't an option you can always
switch vt and kill mayland from there (or edit the config and reload it
once that is supported).