create a dedicated output space
previously i used the smithay `Space`, which provides very little actual
benefit over a relatively simple custom implementation (aside from being
convenient because it was already available) and i had to do most of the
mapping and unmapping logic manually anyway, while being (i think) a lot
more inefficient, and requiring me to do most of the positioning and
mapping stuff in the `WorkspaceManager`, where it was not really
appropriate and where it was cluttering other functions.

the dedicated `OutputSpace` is basically a `Vec` that holds (not unlike
the tiling space) tuples of (Output, Point<i32, Logical>). if you need
the full geometry of the output, that is then calculated with
`output_size`.

when mapping or unmapping an output the current vec of outputs is
drained, and then mapped again. first the outputs that have an explicit
position configured are mapped in the order of their position, then the
rest of the outputs are mapped to the right. if two outputs overlap,
then mayland will still panic for now and that is a problem for future
me to solve, but i think the output space makes any sort of overlap
fixing a lot less annoying.