put active_output into the OutputSpace
it makes more sense there as the mapping and unmapping of outputs can
change the active output, and that functionality is already in the new
`OutputSpace`.

this commit actually only moves the `active_output` field into
`outputs.active` and all of the functionality will be moved in a
seperate commit.