move all of the actual active output handling into the OutputSpace
since adding and removing outputs can change the active output, it makes
a lot of sense for that to be handled inside the OutputSpace.

this also makes a cleaner distinction between the `WorkspaceManager`
and the `OutputSpace` by putting (almost) all of the output stuff into
the `OutputSpace` and making the `WorkspaceManager` almost exclusively
handle the workspaces, which makes the code more maintable and easier to
follow imo.