use the active output when a wlr layer doesn't request one
when a wlr layer is created it can request to be mapped to a specific
output, but it can also let the compositor decide the output.

previously this was just the "next" output, i.e. the one that was first
mapped, but it makes the most sense to use the active output, so that
layer shell tools like fuzzel appear on the correct output.