distinguish between output size and output area changing
when the actual output size changes, by (in the future) reloading the
config, then the outputs have to be `reposition`ed to avoid new output
overlaps or new unnecessary gaps forming

if only the output non_exclusive_zone changes when the layer_map changes,
then i do not have to reposition the outputs, and only have to notify
the workspaces, that their working area changed so that the tiling space
can update accordingly.

this also contains a drive-by bugfix, making it so that all workspaces
of a particular output are notified of the working area changing,
instead of only the one that is active on the output.