new system for queueing and redrawing
instead of inserting an idle into the event loop, the output_state for
every output keeps track of its current QueueState and the redraw
happens in the `refresh` (now `refresh_and_redraw`) function.

this has a few advantages over the old system. for one, it directly
couples the previously distinct `queued` and `waiting_for_vblank` fields.
two, it removes an allocation in queue_redraw_all that has been bugging
me for a while, but was impossible to remove because of rust's aliasing
reules. three, it makes queueing significantly simpler and allows me to
put two asserts into the `QueueState` methods, cleaning up the code. and
last it allows me to redraw after refreshing.

it also made it so i could immediately queue a redraw after a vblank if
a redraw was attempted to be queued while waiting for it. this would
also have been possible to do with the old system but the new system
makes it easier and more convenient.