move the mayfig error print into the mayland_config::Error Display impl
i don't know why i didn't do this in the first place as dealing with the
mayland_config::Error::AlreadyPrinted error was a little annoying, so
this removes that variant and just Display prints the
mayland_config::Error.

with this commit the only change between `Config::init` and
`Config::read` is, that `Config::init` returns a default `Config` on a
`FileNotFound` error, while `Config::read` just returns the error as-is.