rename move_cursor => move_pointer
internally i am usually using the term "cursor" to refer to the rendered
image and "pointer" to refer to the wayland interface that is
responsible for sending motion, scroll ("axis") and other related events
to the wayland clients.

since this fn does not affect the cursor (pointer icon) directly and
only affects the wl_pointer interface calling it "move_cursor" is not
entirely correct.

this is also in line with how the wayland spec uses the terms.