move send_frame_callbacks into on_vblank
rename post_repaint into send_frame_callbacks as that is a significantly
more descriptive name and move the call to that into the on_vblank fn,
but only if the output is not immediately queued for redraw.

this also contains a drive-by bugfix where i immediately requeued on
literally every single vblank because i am sometimes kinda stupid.