unconditionally call send_frame_callbacks
the call to send_frame_callbacks is still in on_vblank in udev, but it
is now always called instead of only being when the output was not
immediately queued again.

this fixes a regression caused by 0d3434c, where window resizing became
extremely laggy. there may be another underlying issue to this, but for
now everything seems to work as intended with this fix.