validate the regex with regex-syntax before adding the ^(?:)$
for a little less of a performance loss just for slightly better error
messages, don't fully compile the regex twice, only check for syntax
once. this also means that regexes have to be valid in itself, meaning
that a regex like "/)(/" no longer works, as it itself is not valid, but
previously i only checked the anchored version "/^(?:)()$/", which was
valid even if it should not have been.