automatically reload the config on change
the reload logic is fairly simple: it `stat`s the file every second and
if the `mtime` changes it tries to reload the config.

i have heard that this does not work on nixos because nixos does stupid
things with symlinks, but i don't have nix and i don't really want to
install it just to test this, so i'll fix it sometime else.