enable -Welided-lifetimes-in-paths lint
part of the rust-2018-idioms, it's allow by default but warns about a
"deprecation" of elided lifetimes in return parameters and arguments,
but, as far as i know, those are not yet deprecated and there's a lot of
discussion around how and when and what to implement about it.

even still, i feel like not eliding the lifetime entirely is the better
decision and makes the code more readable, so it's probably not a bad
decision to lint against it.