properly set and reload the xkb_file config option
when the config is reloaded and the user requests a valid xkb_file,
then it will set the keymap to the xkb_file, otherwise it will fallback
to the XkbConfig.

this is handled in a seperate set_xkb_file method, that will check if
xkb_file is Some, try to read and compile that into a keymap and
setting that as the keymap. any error will be logged and it will only
return true if the keymap was successfully set, allowing me fallback
when the return value is false.

while i still can't set the xkb_file in the Mayland constructor, i can
set it in the two State constructors, and not have to deal with it in
the main function. while at it, also rename that fn to load_xkb_file, as
it does not do anything else as of right now.