use mayland_config::Action as the main Action type
instead of converting from mayland_config::Action to
mayland_comm::Action when deserializing the config, convert from
mayland_comm::Action to mayland_config::Action when a dispatch request
is received.

this is currently only done to make the deserialization code a little
cleaner, but may also be used in the future to add keybinds as a config
option, that should not be available over the wire.