implement per-device input config
you could already configure your input devices in the `input` category
in the mayland.mf config file with the `touchpad {}` and the `mouse {}`
keys, but those are for every single device in that category.

with this change you can now in addition to the general input device
config add a different config for specific devices via a
`touchpad [ "<device name>"] {}` or a `mouse [ "<device name>" ] {}` key.
the options specified in the per-device config will take precedence over
the options in the default config, but everything that was not explicitly
set in the will fall back to the general config in the `touchpad {}` and
`mouse {}` keys respectively.

you can list your current devices and their names using `mayctl devices`

this is not yet currently possible for keyboards, as the xkb config
state is global to the compositor and, unlike input devices handled by
libinput, do not have a per-device config.