allow using '+' in keybind mappings
not sure why i didn't do this earlier (note for the reader: lazyness),
but now you can also do "mod+v".

part of the reason i didn't do this earlier was that it didn't actually
have a benefit over the previous "mod v" as it would be the same size,
but with the recent update to mayfig i have gained the ability use the
'+' characters in identifiers without the need for quotes, so this makes
mapping the keybinds significantly cleaner.