implement my own floating space
i am not entirely happy with how the smithay Space handles things, so
let's implement my own.

this is also in part to implement a window cycling action, as i then
also have to save the insertion order of the windows in a seperate Vec
and makes a lot of sense for that to be saved in the same place.

for now most of the handling of how things work is taken directly from
smithay, especially the `window_under` function, but i'll add my own
implementations bit by bit.