implement the window cycle action
the current floating space saves its windows in z-order, with windows on
top at the back. for cycling through windows this is not very useful, as
focussing a window causes the windows shuffle around as raising a window
to the top puts it at the back of the internal vector.

for that reason the floating space now additionally stores the windows
in insertion order, and then cycling through the floating windows in
that order.

at some point in the future i want to have another Direction enum with
Right and Left, and i want to disambiguate betweeen these two, so i
don't want to use Direction for this.