perf: don't update keyboard focus if it didn't change
previously i would try to update the z-order of all of the windows and
their activation state literally every time when i moved the mouse
pointer. but i don't have to do that when i first check if the focus
actually changed. no idea why i didn't do that from the start.