activate windows directly in the Workspace
instead of calling `set_activate` in the respective spaces, call it when
needed inside of the workspace directly. when mapping a new window, the
space is still responsible for setting the activate state, but is now no
longer responsible for unsetting the activate for the other windows.

this is also technically more correct, as previously i would only unset
the activate state of the windows of the space of the window that i was
activating, which would mean that there could be a floating and a tiling
window active at the same time in the same workspace, which is not what
i want.