make resize take a Rectangle instead of just a Size
it's not really necessary for now, but if i ever add support for
xwayland, then i need a full `Rectangle` when configuring.
additionally the change is not actually that hard and it makes it easier
for the future me that implements xwayland, as i then likely have fewer
merge conflicts with other features.