breaking: switch to kebab-case for the mayfig config
i don't really like snake_case for config files, as it makes the text
look weird, as the underscore is "underneath" the letters instead of
being in-line. additionally it makes certain configs a little less clear,
as something being `left-handed` and tap-and-drag is normal english, but
`left_handed` and `tap_and_drag` is not. some configs will then mix case
and do snake_case, unless for `tap-and-drag`, which is imo a lot worse
than being consistent about it.

switching to kebab-case also makes it align a little bit more with the
plus-seperated keybind+case.

this is an extremely invasive change, so i would prefer to do it earlier
rather than later on and now seems like a pretty good time to do that
still.

this does not affect the types from `mayland-comm`, which are still
snake_case.