switch away from tracing-subscriber
tracing-subscriber pulls in a lot of bullshit like lazy_static, a second
version of regex-automata and nu-ansi-term for exactly no reason, so i
do want to avoid it.

the env_logger crate is not really completely to my needs either: it
doesn't allow me to easily write to multiple sources and writing to
journald is hard as well, but it's not that hard to just directly depend
on env_filter and just handroll most of the rest myself.

this also removes the "debug" feature. by having that behind a
compile-time feature flag, i lose a lot of the benefit of having the
ability to get debug logs: i can't just tell someone to "enable the debug
logs," i have to tell them to recompile mayland first, which is a lot
more of a user burden than just being able to enable it at runtime. this
commit removes the debug flag entirely, but i do intend to bring it back
in the next commit.

similar to the "debug" feature this commit also removes the "trace"
feature, as having to recompile mayland with it is annoying for the exact
same reasons. unlike the "debug" flag, i don't think the trace logs are
actually really useful for like. anything at all, so i don't really have
a reason to enable it again.

additionally you can enable manually enable trace logs again by
specifying a custom RUST_LOG flag, but unlike the previous "trace"
feature, that doesn't split the trace logs into a seperate `.trace.log`
and regular `.log` file. if i ever do need trace logs, then i can add
some way to have trace logs again, but for now i don't need it.