add debug flag to mayland
previously to get a debug log you would have to recompile mayland with
the debug feature, but i think it is very useful to be able to enable
debug logs without recompiling.

i know that that could already be enabled with a `RUST_LOG=debug`
environment variable, but having that as a `--debug` flag does help with
the discoverability and it doesn't like hurt either.