prevent creation of negative size in tiling layout
while it can be triggered with a very high setting for tiling gaps and
border and a small display, it is mostly a theoretical bug (and crash
with debug assertions enabled). however, with a future dwindle layout,
this can be hit in "real" use cases by just spamming window creation.